Collapse _phase4_stubs + phase4 shim into a single phase4 module
The previous round split the Phase-4 surface into two files: a private
_phase4_stubs.py with the umbrella stubs and a phase4.py shim that
re-exported the four working halves. That was clumsy — the umbrella
stubs are public API too (they document the deferred surface and raise
NotImplementedError until the implementation lands), so there is no
reason for them to hide behind an underscore.

Renamed _phase4_stubs.py → phase4.py and deleted the shim. Updated
import sites in tests/test_grid_lidar.py, tests/test_phase4_stubs.py,
and tests/test_phase4_coverage_gaps.py to use the canonical
``from digitalrivers.phase4 import …`` path. Dropped the duplicate
``as stubs`` alias in test_phase4_stubs.py — every reference now uses
``phase4`` directly.

523 unit tests + 4 phase4 doctests still pass.
