+++ src/digitalrivers/phase4.py
@@ -1,35 +1,540 @@
-"""Public surface for Phase 4 helpers (P28-P35).
+"""Phase 4 API surfaces — deferred implementations.
 
-The actual implementations live in ``digitalrivers._phase4_stubs`` alongside
-the umbrella NotImplementedError stubs for the genuinely-deferred halves
-(Dask graph composition, Zarr / S3 / GCS factories, full PDAL pipeline,
-multigrid ANUDEM, edge-flip mesh refinement). This module re-exports the
-working halves under stable, non-underscore-prefixed names so downstream
-callers don't have to import a private path.
+Phase 4 of the digital-rivers roadmap addresses scalability and the more
+research-grade conditioning operations. Each task here is L-effort
+(continental-scale infrastructure: Dask chunking, COG / Zarr I/O, ANUDEM's
+biharmonic + multigrid + spline solver, PDAL LiDAR pipelines,
+post-export mesh-quality refinement, topobathy fusion). None of these can be
+shipped at production quality in a single commit; this module ships the
+*public API surface* so downstream callers can wire against the eventual
+implementations today.
 
-* :func:`tile_windows` — chunked-iteration helper (P30 partial).
-* :func:`write_cog` — Cloud-Optimised GeoTIFF writer (P31 partial).
-* :func:`grid_lidar_points` — LiDAR point cloud → DEM gridding
-  (P34 partial).
-* :func:`topobathy_fusion` — bathymetric DEM fusion (P35).
-
-The umbrella stubs (``dask_backend``, ``cloud_io``, ``anudem_solver``,
-``mesh_quality_optimise``, ``pdal_lidar_pipeline``, ``native_cotat_upscale``,
-``native_ihu_upscale``) remain at ``digitalrivers._phase4_stubs`` because
-they raise ``NotImplementedError`` and are documented as deferred.
+Every function below raises ``NotImplementedError`` with a citation pointing
+at the spec / reference paper. Replace the bodies as each task is properly
+implemented.
 """
 from __future__ import annotations
 
-from digitalrivers._phase4_stubs import (
-    grid_lidar_points,
-    tile_windows,
-    topobathy_fusion,
-    write_cog,
-)
 
-__all__ = [
-    "grid_lidar_points",
-    "tile_windows",
-    "topobathy_fusion",
-    "write_cog",
-]
+def native_cotat_upscale(*args, **kwargs):
+    """Native COTAT / EAM / DMM upscalers (P28).
+
+    The Phase 2 P18 implementation ships COTAT against the public API
+    ``FlowDirection.upscale(method='cotat')``. Phase 4 P28 replaces the
+    pure-Python loop with a Numba-jit native kernel and adds EAM (Olivera
+    2002) and DMM. Effort: L.
+
+    References:
+        Reed S. M. (2003). "Deriving flow directions for coarse-resolution
+        (1-4 km) gridded hydrologic modeling." WRR 39(9):1238.
+        Olivera F. et al. (2002). "Extracting low-resolution river networks
+        from high-resolution digital elevation models." WRR 38(11):1231.
+    """
+    raise NotImplementedError(
+        "Native COTAT/EAM/DMM upscalers (P28) deferred. Phase 2 P18 ships "
+        "a working pure-Python COTAT via FlowDirection.upscale("
+        "method='cotat')."
+    )
+
+
+def native_ihu_upscale(*args, **kwargs):
+    """Native IHU (P29, Eilander 2021) — now implemented via P19.
+
+    The hill-climbing swap-search engine lives in
+    ``digitalrivers._ihu.ihu_upscale`` and is exposed through
+    ``FlowDirection.upscale_ihu(...)`` and ``FlowDirection.upscale(
+    method="ihu", ...)``. This umbrella stub is kept for API-discovery
+    symmetry with the other P28-P35 entry points; it points callers at
+    the real implementation.
+
+    Implementation: greedy hill-climbing on a global drainage-area-error
+    metric. Starts from a COTAT initial network; for each iteration,
+    each coarse cell tries every alternative outlet in turn and accepts
+    the first swap that reduces the global metric. Converges when no
+    single-cell swap improves; returns ``converged`` in the metrics
+    dict.
+
+    Performance: pure Python. Works on small/medium DEMs (thousands of
+    cells) in seconds; a Numba port is a follow-up. For continental
+    DEMs, consider the pyflwdir vendor path until then.
+
+    References:
+        Eilander D. et al. (2021). "A hydrography upscaling method for
+        scale-invariant parametrization of distributed hydrological
+        models." HESS 25(9):5287-5313.
+    """
+    raise NotImplementedError(
+        "native_ihu_upscale umbrella API: use "
+        "FlowDirection.upscale_ihu(scale_factor, accumulation, dem, "
+        "max_iter, report) or FlowDirection.upscale(scale_factor, "
+        "method='ihu', accumulation, dem) — both wire through the "
+        "working IHU engine in digitalrivers._ihu."
+    )
+
+
+def tile_windows(
+    dataset,
+    tile_rows: int = 1024,
+    tile_cols: int = 1024,
+    overlap: int = 0,
+):
+    """Iterate ``(row_off, col_off, n_rows, n_cols)`` tile windows over a Dataset.
+
+    Partial implementation of P30 — Dask-style chunked iteration without
+    the Dask dependency. Yields one window per tile so callers can stream
+    a continental DEM through any per-tile algorithm without ever
+    materialising the full raster in memory. Each window is a
+    GDAL-compatible ``(xoff, yoff, xsize, ysize)`` quadruple ready to
+    pass into ``Dataset.read_array(window=...)``.
+
+    Tile size defaults match the COG / Cloud-Optimised GeoTIFF spec
+    (512×512 or 1024×1024 internal tiles).
+
+    Args:
+        dataset: A pyramids ``Dataset`` (or subclass).
+        tile_rows: Tile height in cells. Defaults to 1024.
+        tile_cols: Tile width in cells. Defaults to 1024.
+        overlap: Cells of overlap between adjacent tiles. Useful for
+            algorithms that need neighbour context (slopes, flow
+            direction, dilations). Default 0.
+
+    Yields:
+        ``(row_off, col_off, n_rows, n_cols)`` int tuples in row-major
+        order. Edge tiles are clipped to the dataset bounds.
+
+    Examples:
+        - Iterate a 5x5 dataset in 3x3 tiles with no overlap:
+
+            >>> import numpy as np
+            >>> from pyramids.dataset import Dataset
+            >>> from digitalrivers.phase4 import tile_windows
+            >>> ds = Dataset.create_from_array(
+            ...     np.zeros((5, 5), dtype=np.float32),
+            ...     top_left_corner=(0, 0), cell_size=1.0, epsg=4326,
+            ... )
+            >>> windows = list(tile_windows(ds, tile_rows=3, tile_cols=3))
+            >>> [(w[0], w[1], w[2], w[3]) for w in windows]
+            [(0, 0, 3, 3), (0, 3, 3, 2), (3, 0, 2, 3), (3, 3, 2, 2)]
+    """
+    rows = dataset.rows
+    cols = dataset.columns
+    if tile_rows <= 0 or tile_cols <= 0:
+        raise ValueError("tile_rows and tile_cols must be positive")
+    if overlap < 0:
+        raise ValueError("overlap must be >= 0")
+    step_r = max(1, tile_rows - overlap)
+    step_c = max(1, tile_cols - overlap)
+    for r_off in range(0, rows, step_r):
+        n_r = min(tile_rows, rows - r_off)
+        if n_r <= 0:
+            break
+        for c_off in range(0, cols, step_c):
+            n_c = min(tile_cols, cols - c_off)
+            if n_c <= 0:
+                break
+            yield (r_off, c_off, n_r, n_c)
+
+
+def dask_backend(*args, **kwargs):
+    """Dask / chunked-tile backend for continental DEMs (P30) — umbrella.
+
+    Full Dask-graph integration remains deferred. The chunked-iteration
+    half ships as :func:`tile_windows` — callers process continental DEMs
+    by looping ``for win in tile_windows(ds): chunk = ds.read_array(window=win)``
+    without loading the full mosaic in memory.
+
+    References:
+        Dask documentation: https://docs.dask.org/
+        rioxarray chunked I/O.
+    """
+    raise NotImplementedError(
+        "dask_backend umbrella API (P30) deferred. Use "
+        "digitalrivers.phase4.tile_windows for per-tile streaming."
+    )
+
+
+def write_cog(dataset, path: str, compress: str = "deflate") -> str:
+    """Cloud-Optimised GeoTIFF writer (P31 partial).
+
+    Writes a pyramids ``Dataset`` to a COG file using GDAL's built-in COG
+    driver. COG is the standard cloud-native format for raster data:
+    internally tiled, internally overviewed, and indexable by HTTP range
+    requests — the foundation of every modern STAC-based pipeline.
+
+    The full P31 scope also includes Zarr writers and S3 / GCS read
+    factories; those remain deferred.
+
+    Args:
+        dataset: Any ``pyramids.Dataset`` (or subclass — DEM,
+            FlowDirection, Accumulation, etc.).
+        path: Output ``.tif`` path.
+        compress: GDAL compression option (``"deflate"`` default,
+            ``"lzw"``, ``"zstd"``, ``"none"``).
+
+    Returns:
+        The output path on success.
+
+    Raises:
+        RuntimeError: If GDAL's COG driver fails (older GDAL builds may
+            need ``"GTIFF"`` with manual COG options instead).
+
+    Examples:
+        - Write a 5x5 DEM as a COG:
+
+            >>> import numpy as np
+            >>> from pyramids.dataset import Dataset
+            >>> from digitalrivers.phase4 import write_cog
+            >>> import tempfile, os
+            >>> arr = np.arange(25, dtype=np.float32).reshape(5, 5)
+            >>> ds = Dataset.create_from_array(
+            ...     arr, top_left_corner=(0, 0), cell_size=1.0,
+            ...     epsg=4326,
+            ... )
+            >>> tmp = tempfile.NamedTemporaryFile(suffix='.tif', delete=False)
+            >>> _ = tmp.close()
+            >>> out_path = write_cog(ds, tmp.name)
+            >>> os.path.exists(out_path)
+            True
+            >>> os.unlink(out_path)
+    """
+    from osgeo import gdal
+
+    src = dataset.raster
+    driver = gdal.GetDriverByName("COG")
+    if driver is None:
+        raise RuntimeError(
+            "GDAL COG driver not available; upgrade GDAL >= 3.1 or write "
+            "via GTIFF with TILED=YES + COPY_SRC_OVERVIEWS=YES manually."
+        )
+    options = [f"COMPRESS={compress.upper()}"]
+    out = driver.CreateCopy(path, src, 0, options)
+    if out is None:
+        raise RuntimeError(f"GDAL COG driver failed to write {path}")
+    out = None  # flush / close
+    return path
+
+
+def cloud_io(*args, **kwargs):
+    """Cloud-optimised raster I/O (P31): COG / Zarr / S3 / GCS — deferred.
+
+    The COG write half is shipped under :func:`write_cog`. Zarr writers
+    and S3 / GCS read factories remain deferred pending a follow-up PR.
+    """
+    raise NotImplementedError(
+        "cloud_io umbrella API (P31) deferred. The COG write half is "
+        "available via digitalrivers.phase4.write_cog."
+    )
+
+
+def anudem_solver(*args, **kwargs):
+    """Full ANUDEM (P32, Hutchinson 1989) — partial via P25.
+
+    DEM.anudem_interpolate(method="biharmonic") solves Δ²z = 0 via
+    alternating Laplacian sweeps, providing the biharmonic objective
+    half of Hutchinson 1989 (without multigrid acceleration or the
+    drainage-enforcement constraints). For drainage enforcement combine
+    with DEM.burn_streams before or after.
+
+    The full multigrid + spline + drainage-enforcement solver remains
+    a follow-up.
+
+    References:
+        Hutchinson M. F. (1989). "A new procedure for gridding elevation
+        and stream line data with automatic removal of spurious pits."
+        Journal of Hydrology 106(3-4):211-232.
+    """
+    raise NotImplementedError(
+        "anudem_solver umbrella API: use "
+        "DEM.anudem_interpolate(method='biharmonic') for the biharmonic "
+        "core, or method='laplacian' (default) for a faster harmonic "
+        "extension. Drainage enforcement: chain with DEM.burn_streams."
+    )
+
+
+def mesh_quality_optimise(*args, **kwargs):
+    """Post-export mesh quality optimisation (P33).
+
+    Operates on the meshes produced by Phase 3 P26 exporters; applies
+    Laplacian smoothing, edge flips, and refinement around breaklines to
+    improve element aspect ratios. The Laplacian-smoothing half is
+    available via :class:`digitalrivers.mesh.Mesh` and
+    :meth:`Mesh.laplacian_smooth`; edge flips and breakline-aware
+    refinement remain deferred.
+
+    References:
+        Persson P.-O., Strang G. (2004). "A simple mesh generator in
+        MATLAB." SIAM Review 46(2):329-345.
+    """
+    raise NotImplementedError(
+        "mesh_quality_optimise umbrella API: use digitalrivers.mesh.Mesh "
+        "and Mesh.laplacian_smooth() for the smoothing half. Edge flips "
+        "and breakline-aware refinement remain deferred."
+    )
+
+
+def grid_lidar_points(
+    xs,
+    ys,
+    zs,
+    cell_size: float,
+    bounds=None,
+    aggregate: str = "min",
+    epsg: int = 4326,
+):
+    """Grid a LiDAR point cloud to a DEM (P34 partial).
+
+    Pragmatic LiDAR-to-DEM step that operates on raw ``(x, y, z)`` arrays
+    — useful when the caller has read LAS / LAZ externally (via ``laspy``,
+    ``pylas``, or PDAL) and wants a gridded surface. The full PDAL pipeline
+    (read + classify + ground-filter + grid + condition) remains deferred.
+
+    For each cell, aggregates the z values of every point that lands in
+    it. ``aggregate`` controls the aggregation: ``"min"`` (default — the
+    canonical bare-earth choice for first-return LiDAR), ``"max"``,
+    ``"mean"``, or ``"median"``. Cells with no points receive the dataset
+    no-data sentinel.
+
+    Args:
+        xs / ys / zs: 1-D arrays of point coordinates.
+        cell_size: output cell side length in map units (must match the
+            CRS).
+        bounds: ``(x_min, y_min, x_max, y_max)`` to clip the grid to. If
+            ``None``, the input points' bounding box is used.
+        aggregate: ``"min"`` (default), ``"max"``, ``"mean"``, ``"median"``.
+        epsg: EPSG code of the input coordinates.
+
+    Returns:
+        A pyramids ``Dataset`` of the gridded surface.
+
+    Raises:
+        ValueError: For mismatched input lengths or unknown ``aggregate``.
+
+    Examples:
+        - Bucket four points into a 2x1 grid with min aggregation:
+
+            >>> import numpy as np
+            >>> from digitalrivers.phase4 import grid_lidar_points
+            >>> xs = np.array([0.1, 0.2, 1.1])
+            >>> ys = np.array([0.1, 0.2, 0.1])
+            >>> zs = np.array([5.0, 2.0, 4.0])
+            >>> ds = grid_lidar_points(
+            ...     xs, ys, zs, cell_size=1.0, bounds=(0.0, 0.0, 2.0, 1.0),
+            ...     aggregate="min", epsg=3857,
+            ... )
+            >>> ds.read_array().tolist()
+            [[2.0, 4.0]]
+
+        - Mean aggregation averages every point that lands in a cell:
+
+            >>> import numpy as np
+            >>> from digitalrivers.phase4 import grid_lidar_points
+            >>> xs = np.array([0.1, 0.2])
+            >>> ys = np.array([0.1, 0.2])
+            >>> zs = np.array([4.0, 6.0])
+            >>> ds = grid_lidar_points(
+            ...     xs, ys, zs, cell_size=1.0, bounds=(0.0, 0.0, 1.0, 1.0),
+            ...     aggregate="mean", epsg=3857,
+            ... )
+            >>> float(ds.read_array()[0, 0])
+            5.0
+
+        - Empty cells receive the dataset's no-data sentinel (-9999.0):
+
+            >>> import numpy as np
+            >>> from digitalrivers.phase4 import grid_lidar_points
+            >>> ds = grid_lidar_points(
+            ...     np.array([0.5]), np.array([0.5]), np.array([3.0]),
+            ...     cell_size=1.0, bounds=(0.0, 0.0, 2.0, 2.0),
+            ...     aggregate="min",
+            ... )
+            >>> float(ds.no_data_value[0])
+            -9999.0
+            >>> int((ds.read_array() == -9999.0).sum())
+            3
+    """
+    import numpy as np
+    from pyramids.dataset import Dataset
+
+    xs = np.asarray(xs, dtype=np.float64)
+    ys = np.asarray(ys, dtype=np.float64)
+    zs = np.asarray(zs, dtype=np.float64)
+    if not (len(xs) == len(ys) == len(zs)):
+        raise ValueError(
+            f"xs, ys, zs must have the same length; got {len(xs)}, "
+            f"{len(ys)}, {len(zs)}"
+        )
+    if aggregate not in ("min", "max", "mean", "median"):
+        raise ValueError(
+            f"aggregate must be one of 'min','max','mean','median'; "
+            f"got {aggregate!r}"
+        )
+    if bounds is None:
+        x_min, y_min = float(xs.min()), float(ys.min())
+        x_max, y_max = float(xs.max()), float(ys.max())
+    else:
+        x_min, y_min, x_max, y_max = bounds
+
+    cols = int(np.ceil((x_max - x_min) / cell_size))
+    rows = int(np.ceil((y_max - y_min) / cell_size))
+    # Map each point to its (row, col).
+    col_idx = np.clip(((xs - x_min) / cell_size).astype(np.int64), 0, cols - 1)
+    row_idx = np.clip(
+        ((y_max - ys) / cell_size).astype(np.int64), 0, rows - 1
+    )
+
+    nodata = -9999.0
+    # ``mean`` uses ``np.add.at`` for an O(N_points) reduction; ``min`` /
+    # ``max`` use ``np.minimum.at`` / ``np.maximum.at`` for the same.
+    # ``median`` still requires per-cell bucketing because there is no
+    # closed-form running median in NumPy.
+    if aggregate == "min":
+        out = np.full((rows, cols), np.inf, dtype=np.float64)
+        np.minimum.at(out, (row_idx, col_idx), zs)
+        out[~np.isfinite(out)] = nodata
+    elif aggregate == "max":
+        out = np.full((rows, cols), -np.inf, dtype=np.float64)
+        np.maximum.at(out, (row_idx, col_idx), zs)
+        out[~np.isfinite(out)] = nodata
+    elif aggregate == "mean":
+        sums = np.zeros((rows, cols), dtype=np.float64)
+        counts = np.zeros((rows, cols), dtype=np.int64)
+        np.add.at(sums, (row_idx, col_idx), zs)
+        np.add.at(counts, (row_idx, col_idx), 1)
+        with np.errstate(invalid="ignore", divide="ignore"):
+            out = np.where(counts > 0, sums / counts, nodata)
+    else:  # median — per-cell bucketing
+        buckets: dict[tuple[int, int], list[float]] = {}
+        for r, c, z in zip(row_idx, col_idx, zs):
+            buckets.setdefault((int(r), int(c)), []).append(float(z))
+        out = np.full((rows, cols), nodata, dtype=np.float64)
+        for (r, c), vals in buckets.items():
+            out[r, c] = float(np.median(np.asarray(vals, dtype=np.float64)))
+
+    geo = (x_min, cell_size, 0.0, y_max, 0.0, -cell_size)
+    return Dataset.create_from_array(
+        out.astype(np.float32, copy=False),
+        geo=geo, epsg=epsg, no_data_value=nodata,
+    )
+
+
+def pdal_lidar_pipeline(*args, **kwargs):
+    """Full PDAL pipeline (P34) — umbrella.
+
+    The point-cloud-to-DEM gridding half ships as :func:`grid_lidar_points`
+    and works on raw ``(x, y, z)`` arrays. The full PDAL pipeline (LAS /
+    LAZ read + ground classification + filtering + grid + Phase 1-3
+    conditioning chain) remains deferred pending the PDAL conda-forge
+    dependency.
+    """
+    raise NotImplementedError(
+        "pdal_lidar_pipeline umbrella API (P34) deferred. The gridding "
+        "half is available via digitalrivers.phase4.grid_lidar_points"
+        "; read LAS files externally with laspy / pylas / PDAL and pass "
+        "the resulting arrays into grid_lidar_points."
+    )
+
+
+def topobathy_fusion(
+    topo,
+    bathy,
+    shoreline_elev: float = 0.0,
+    blend: str = "max",
+):
+    """Bathymetric DEM fusion (P35).
+
+    Fuses a topographic DEM and a bathymetric DEM into a single
+    hydrographic surface. Both inputs must be aligned (same shape,
+    geotransform, CRS). The shoreline is the contour at
+    ``shoreline_elev`` (default 0 — mean sea level).
+
+    Blend modes:
+
+    * ``"max"`` (default): per-cell maximum of the two DEMs. Topo wins
+      above the shoreline, bathy below — the canonical conservative
+      choice when the two DEMs disagree across the shoreline (NOAA
+      ETOPO uses this).
+    * ``"min"``: per-cell minimum — the pessimistic-bathymetry choice
+      for flood inundation studies where you want to assume the deeper
+      of two conflicting surveys.
+    * ``"topo_above"``: pick topo where ``topo >= shoreline_elev``,
+      bathy elsewhere. Sharp transition at the shoreline; preferred when
+      the topo DEM is known accurate at the coastline.
+    * ``"bathy_below"``: pick bathy where ``bathy <= shoreline_elev``,
+      topo elsewhere. Mirror of the above.
+
+    Args:
+        topo: Topographic ``Dataset`` (DEM subclass acceptable).
+        bathy: Bathymetric ``Dataset`` aligned to ``topo``.
+        shoreline_elev: Elevation defining the shoreline contour.
+            Default 0.0 (MSL).
+        blend: ``"max"`` (default), ``"min"``, ``"topo_above"``, or
+            ``"bathy_below"``.
+
+    Returns:
+        ``Dataset`` of the fused surface.
+
+    Raises:
+        ValueError: If shapes mismatch or ``blend`` is unknown.
+
+    Examples:
+        - The ``"min"`` blend picks the cell-by-cell minimum — useful as a
+          pessimistic-bathymetry baseline:
+
+            >>> import numpy as np
+            >>> from pyramids.dataset import Dataset
+            >>> from digitalrivers.phase4 import topobathy_fusion
+            >>> topo = Dataset.create_from_array(
+            ...     np.array([[5.0, -1.0]], dtype=np.float32),
+            ...     top_left_corner=(0, 0), cell_size=1.0, epsg=4326,
+            ... )
+            >>> bathy = Dataset.create_from_array(
+            ...     np.array([[-3.0, -5.0]], dtype=np.float32),
+            ...     top_left_corner=(0, 0), cell_size=1.0, epsg=4326,
+            ... )
+            >>> fused = topobathy_fusion(topo, bathy, blend="min")
+            >>> fused.read_array().tolist()
+            [[-3.0, -5.0]]
+
+    References:
+        Eakins B. W., Grothe P. R. (2014). "Challenges in building coastal
+        digital elevation models." Journal of Coastal Research 30(5).
+    """
+    import numpy as np
+    from pyramids.dataset import Dataset
+
+    if blend not in ("max", "min", "topo_above", "bathy_below"):
+        raise ValueError(
+            f"blend must be one of 'max', 'min', 'topo_above', "
+            f"'bathy_below'; got {blend!r}"
+        )
+
+    topo_arr = topo.read_array().astype(np.float64, copy=False)
+    bathy_arr = bathy.read_array().astype(np.float64, copy=False)
+    if topo_arr.shape != bathy_arr.shape:
+        raise ValueError(
+            f"topo shape {topo_arr.shape} != bathy shape {bathy_arr.shape}"
+        )
+
+    # Replace no-data with NaN for blending so np.fmax / np.where handle gaps.
+    topo_no_val = topo.no_data_value[0] if topo.no_data_value else None
+    bathy_no_val = bathy.no_data_value[0] if bathy.no_data_value else None
+    if topo_no_val is not None:
+        topo_arr = np.where(topo_arr == topo_no_val, np.nan, topo_arr)
+    if bathy_no_val is not None:
+        bathy_arr = np.where(bathy_arr == bathy_no_val, np.nan, bathy_arr)
+
+    if blend == "max":
+        fused = np.fmax(topo_arr, bathy_arr)
+    elif blend == "min":
+        fused = np.fmin(topo_arr, bathy_arr)
+    elif blend == "topo_above":
+        fused = np.where(topo_arr >= shoreline_elev, topo_arr, bathy_arr)
+    else:  # bathy_below
+        fused = np.where(bathy_arr <= shoreline_elev, bathy_arr, topo_arr)
+
+    out_no_val = topo_no_val if topo_no_val is not None else -9999.0
+    fused = np.where(np.isnan(fused), out_no_val, fused)
+    return Dataset.create_from_array(
+        fused.astype(np.float32, copy=False),
+        geo=topo.geotransform, epsg=topo.epsg, no_data_value=out_no_val,
+    )
